requirements.txt
Flask==3.0.3
APScheduler==3.10.4
plexapi
qbittorrent-api
requests
waitress
python-dotenv==1.0.1
